Push all the tags for the built image
Podman push only pushes the latest tag provided as argument to the command. So if an image is tagged multiple times, the command has to be executed multiple times. The actual image only gets pushed once, as the registry detects that the image's content is the same, but makes an entry for each tag.